+++ src/muleengine/mesh_creator/cone.rs
@@ -65,9 +65,9 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
         );
 
         mesh.add_face(
-            mesh.number_of_vertices() - 1,
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 3,
+            (mesh.number_of_vertices() - 1) as u32,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 3) as u32,
         );
     }
 
@@ -123,9 +123,9 @@ pub fn create(radius: f32, height: f32, resolution: i32) -> Mesh {
         );
 
         mesh.add_face(
-            mesh.number_of_vertices() - 1,
-            mesh.number_of_vertices() - 2,
-            mesh.number_of_vertices() - 3,
+            (mesh.number_of_vertices() - 1) as u32,
+            (mesh.number_of_vertices() - 2) as u32,
+            (mesh.number_of_vertices() - 3) as u32,
         );
     }
